refactor: Migration des types pour les composants de groupes et idées cadeaux
